revert android changes
