+++ shell/platform/android/BUILD.gn
@@ -70,8 +70,6 @@ source_set("flutter_shell_native_src") {
     "android_context_gl_impeller.h",
     "android_context_gl_skia.cc",
     "android_context_gl_skia.h",
-    "android_context_vulkan_impeller.cc",
-    "android_context_vulkan_impeller.h",
     "android_display.cc",
     "android_display.h",
     "android_egl_surface.cc",
