--- shell/platform/android/android_context_vulkan_impeller.h
@@ -1,35 +1,0 @@
-// Copyright 2013 The Flutter Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-#ifndef FLUTTER_SHELL_PLATFORM_ANDROID_ANDROID_CONTEXT_VULKAN_IMPELLER_H_
-#define FLUTTER_SHELL_PLATFORM_ANDROID_ANDROID_CONTEXT_VULKAN_IMPELLER_H_
-
-#include "flutter/fml/concurrent_message_loop.h"
-#include "flutter/fml/macros.h"
-#include "flutter/impeller/toolkit/egl/display.h"
-#include "flutter/shell/platform/android/context/android_context.h"
-#include "flutter/vulkan/procs/vulkan_proc_table.h"
-
-namespace flutter {
-
-class AndroidContextVulkanImpeller : public AndroidContext {
- public:
-  explicit AndroidContextVulkanImpeller(bool enable_validation);
-
-  ~AndroidContextVulkanImpeller();
-
-  // |AndroidContext|
-  bool IsValid() const override;
-
- private:
-  fml::RefPtr<vulkan::VulkanProcTable> proc_table_;
-  std::shared_ptr<fml::ConcurrentMessageLoop> workers_;
-  bool is_valid_ = false;
-
-  FML_DISALLOW_COPY_AND_ASSIGN(AndroidContextVulkanImpeller);
-};
-
-}  // namespace flutter
-
-#endif  // FLUTTER_SHELL_PLATFORM_ANDROID_ANDROID_CONTEXT_VULKAN_IMPELLER_H_
