+++ shell/platform/android/android_surface_gl_skia.cc
@@ -21,13 +21,14 @@ constexpr char kEmulatorRendererPrefix[] =
 }  // anonymous namespace
 
 AndroidSurfaceGLSkia::AndroidSurfaceGLSkia(
-    const std::shared_ptr<AndroidContextGLSkia>& android_context)
-    : android_context_(android_context),
+    const std::shared_ptr<AndroidContext>& android_context,
+    const std::shared_ptr<PlatformViewAndroidJNI>& jni_facade)
+    : AndroidSurface(android_context),
       native_window_(nullptr),
       onscreen_surface_(nullptr),
       offscreen_surface_(nullptr) {
   // Acquire the offscreen surface.
-  offscreen_surface_ = android_context_->CreateOffscreenSurface();
+  offscreen_surface_ = GLContextPtr()->CreateOffscreenSurface();
   if (!offscreen_surface_->IsValid()) {
     offscreen_surface_ = nullptr;
   }
@@ -39,12 +40,12 @@ void AndroidSurfaceGLSkia::TeardownOnScreenContext() {
   // When the onscreen surface is destroyed, the context and the surface
   // instance should be deleted. Issue:
   // https://github.com/flutter/flutter/issues/64414
-  android_context_->ClearCurrent();
+  GLContextPtr()->ClearCurrent();
   onscreen_surface_ = nullptr;
 }
 
 bool AndroidSurfaceGLSkia::IsValid() const {
-  return offscreen_surface_ && android_context_->IsValid();
+  return offscreen_surface_ && GLContextPtr()->IsValid();
 }
 
 std::unique_ptr<Surface> AndroidSurfaceGLSkia::CreateGPUSurface(
@@ -54,10 +55,10 @@ std::unique_ptr<Surface> AndroidSurfaceGLSkia::CreateGPUSurface(
                                               true);
   } else {
     sk_sp<GrDirectContext> main_skia_context =
-        android_context_->GetMainSkiaContext();
+        GLContextPtr()->GetMainSkiaContext();
     if (!main_skia_context) {
       main_skia_context = GPUSurfaceGLSkia::MakeGLContext(this);
-      android_context_->SetMainSkiaContext(main_skia_context);
+      GLContextPtr()->SetMainSkiaContext(main_skia_context);
     }
     return std::make_unique<GPUSurfaceGLSkia>(main_skia_context, this, true);
   }
@@ -72,12 +73,12 @@ bool AndroidSurfaceGLSkia::OnScreenSurfaceResize(const SkISize& size) {
     return true;
   }
 
-  android_context_->ClearCurrent();
+  GLContextPtr()->ClearCurrent();
 
   // Ensure the destructor is called since it destroys the `EGLSurface` before
   // creating a new onscreen surface.
   onscreen_surface_ = nullptr;
-  onscreen_surface_ = android_context_->CreateOnscreenSurface(native_window_);
+  onscreen_surface_ = GLContextPtr()->CreateOnscreenSurface(native_window_);
   if (!onscreen_surface_->IsValid()) {
     FML_LOG(ERROR) << "Unable to create EGL window surface on resize.";
     return false;
@@ -108,7 +109,7 @@ bool AndroidSurfaceGLSkia::SetNativeWindow(
   // creating a new onscreen surface.
   onscreen_surface_ = nullptr;
   // Create the onscreen surface.
-  onscreen_surface_ = android_context_->CreateOnscreenSurface(window);
+  onscreen_surface_ = GLContextPtr()->CreateOnscreenSurface(window);
   if (!onscreen_surface_->IsValid()) {
     return false;
   }
@@ -126,7 +127,7 @@ std::unique_ptr<GLContextResult> AndroidSurfaceGLSkia::GLContextMakeCurrent() {
 
 bool AndroidSurfaceGLSkia::GLContextClearCurrent() {
   FML_DCHECK(IsValid());
-  return android_context_->ClearCurrent();
+  return GLContextPtr()->ClearCurrent();
 }
 
 SurfaceFrame::FramebufferInfo AndroidSurfaceGLSkia::GLContextFramebufferInfo()
@@ -185,7 +186,7 @@ sk_sp<const GrGLInterface> AndroidSurfaceGLSkia::GetGLInterface() const {
       reinterpret_cast<const char*>(glGetString(GL_RENDERER));
   if (gl_renderer && strncmp(gl_renderer, kEmulatorRendererPrefix,
                              strlen(kEmulatorRendererPrefix)) == 0) {
-    EGLContext new_context = android_context_->CreateNewContext();
+    EGLContext new_context = GLContextPtr()->CreateNewContext();
     if (new_context != EGL_NO_CONTEXT) {
       EGLContext old_context = eglGetCurrentContext();
       EGLDisplay display = eglGetCurrentDisplay();
@@ -204,15 +205,19 @@ sk_sp<const GrGLInterface> AndroidSurfaceGLSkia::GetGLInterface() const {
   return GPUSurfaceGLDelegate::GetGLInterface();
 }
 
+AndroidContextGLSkia* AndroidSurfaceGLSkia::GLContextPtr() const {
+  return reinterpret_cast<AndroidContextGLSkia*>(android_context_.get());
+}
+
 std::unique_ptr<Surface> AndroidSurfaceGLSkia::CreateSnapshotSurface() {
   if (!onscreen_surface_ || !onscreen_surface_->IsValid()) {
-    onscreen_surface_ = android_context_->CreatePbufferSurface();
+    onscreen_surface_ = GLContextPtr()->CreatePbufferSurface();
   }
   sk_sp<GrDirectContext> main_skia_context =
-      android_context_->GetMainSkiaContext();
+      GLContextPtr()->GetMainSkiaContext();
   if (!main_skia_context) {
     main_skia_context = GPUSurfaceGLSkia::MakeGLContext(this);
-    android_context_->SetMainSkiaContext(main_skia_context);
+    GLContextPtr()->SetMainSkiaContext(main_skia_context);
   }
 
   return std::make_unique<GPUSurfaceGLSkia>(main_skia_context, this, true);
