+++ shell/platform/android/android_surface_gl_skia.h
@@ -21,8 +21,9 @@ namespace flutter {
 class AndroidSurfaceGLSkia final : public GPUSurfaceGLDelegate,
                                    public AndroidSurface {
  public:
-  explicit AndroidSurfaceGLSkia(
-      const std::shared_ptr<AndroidContextGLSkia>& android_context);
+  AndroidSurfaceGLSkia(
+      const std::shared_ptr<AndroidContext>& android_context,
+      const std::shared_ptr<PlatformViewAndroidJNI>& jni_facade);
 
   ~AndroidSurfaceGLSkia() override;
 
@@ -81,10 +82,15 @@ class AndroidSurfaceGLSkia final : public GPUSurfaceGLDelegate,
   }
 
  private:
-  std::shared_ptr<AndroidContextGLSkia> android_context_;
   fml::RefPtr<AndroidNativeWindow> native_window_;
   std::unique_ptr<AndroidEGLSurface> onscreen_surface_;
   std::unique_ptr<AndroidEGLSurface> offscreen_surface_;
+
+  //----------------------------------------------------------------------------
+  /// @brief      Takes the super class AndroidSurface's AndroidContext and
+  ///             return a raw pointer to an AndroidContextGL.
+  ///
+  AndroidContextGLSkia* GLContextPtr() const;
 
   FML_DISALLOW_COPY_AND_ASSIGN(AndroidSurfaceGLSkia);
 };
