+++ shell/platform/android/android_surface_software.cc
@@ -40,7 +40,10 @@ bool GetSkColorType(int32_t buffer_format,
 
 }  // anonymous namespace
 
-AndroidSurfaceSoftware::AndroidSurfaceSoftware() {
+AndroidSurfaceSoftware::AndroidSurfaceSoftware(
+    const std::shared_ptr<AndroidContext>& android_context,
+    const std::shared_ptr<PlatformViewAndroidJNI>& jni_facade)
+    : AndroidSurface(android_context) {
   GetSkColorType(WINDOW_FORMAT_RGBA_8888, &target_color_type_,
                  &target_alpha_type_);
 }
