+++ shell/platform/android/android_surface_vulkan_impeller.h
@@ -7,16 +7,18 @@
 #include "flutter/fml/concurrent_message_loop.h"
 #include "flutter/fml/macros.h"
 #include "flutter/impeller/renderer/context.h"
-#include "flutter/shell/platform/android/android_context_vulkan_impeller.h"
 #include "flutter/shell/platform/android/surface/android_native_window.h"
 #include "flutter/shell/platform/android/surface/android_surface.h"
+#include "flutter/vulkan/procs/vulkan_proc_table.h"
 
 namespace flutter {
 
 class AndroidSurfaceVulkanImpeller : public AndroidSurface {
  public:
-  explicit AndroidSurfaceVulkanImpeller(
-      const std::shared_ptr<AndroidContextVulkanImpeller>& android_context);
+  AndroidSurfaceVulkanImpeller(
+      const std::shared_ptr<AndroidContext>& android_context,
+      const std::shared_ptr<PlatformViewAndroidJNI>& jni_facade,
+      bool enable_vulkan_validation);
 
   ~AndroidSurfaceVulkanImpeller() override;
 
@@ -46,8 +48,10 @@ class AndroidSurfaceVulkanImpeller : public AndroidSurface {
   bool SetNativeWindow(fml::RefPtr<AndroidNativeWindow> window) override;
 
  private:
-  std::shared_ptr<AndroidContextVulkanImpeller> android_context_;
+  fml::RefPtr<vulkan::VulkanProcTable> proc_table_;
   fml::RefPtr<AndroidNativeWindow> native_window_;
+  std::shared_ptr<fml::ConcurrentMessageLoop> workers_;
+  std::shared_ptr<impeller::Context> impeller_context_;
   bool is_valid_ = false;
 
   FML_DISALLOW_COPY_AND_ASSIGN(AndroidSurfaceVulkanImpeller);
