+++ shell/platform/android/context/android_context.cc
@@ -32,13 +32,4 @@ sk_sp<GrDirectContext> AndroidContext::GetMainSkiaContext() const {
   return main_context_;
 }
 
-std::shared_ptr<impeller::Context> AndroidContext::GetImpellerContext() const {
-  return impeller_context_;
-}
-
-void AndroidContext::SetImpellerContext(
-    const std::shared_ptr<impeller::Context>& context) {
-  impeller_context_ = context;
-}
-
 }  // namespace flutter
