+++ shell/platform/android/context/android_context.h
@@ -7,7 +7,6 @@
 
 #include "flutter/fml/macros.h"
 #include "flutter/fml/task_runner.h"
-#include "flutter/impeller/renderer/context.h"
 #include "third_party/skia/include/gpu/GrDirectContext.h"
 
 namespace flutter {
@@ -15,7 +14,6 @@ namespace flutter {
 enum class AndroidRenderingAPI {
   kSoftware,
   kOpenGLES,
-  kVulkan,
 };
 
 //------------------------------------------------------------------------------
@@ -56,24 +54,11 @@ class AndroidContext {
   ///
   sk_sp<GrDirectContext> GetMainSkiaContext() const;
 
-  //----------------------------------------------------------------------------
-  /// @brief      Accessor for the Impeller context associated with
-  ///             AndroidSurfaces and the raster thread.
-  ///
-  std::shared_ptr<impeller::Context> GetImpellerContext() const;
-
- protected:
-  /// Intended to be called from a subclass constructor after setup work for the
-  /// context has completed.
-  void SetImpellerContext(const std::shared_ptr<impeller::Context>& context);
-
  private:
   const AndroidRenderingAPI rendering_api_;
 
   // This is the Skia context used for on-screen rendering.
   sk_sp<GrDirectContext> main_context_;
-
-  std::shared_ptr<impeller::Context> impeller_context_;
 
   FML_DISALLOW_COPY_AND_ASSIGN(AndroidContext);
 };
