+++ shell/platform/android/external_view_embedder/surface_pool_unittests.cc
@@ -50,9 +50,10 @@ TEST(SurfacePool, GetLayerAllocateOneLayer) {
           ByMove(std::make_unique<PlatformViewAndroidJNI::OverlayMetadata>(
               0, window))));
 
-  auto surface_factory =
-      std::make_shared<TestAndroidSurfaceFactory>([gr_context, window]() {
-        auto android_surface_mock = std::make_unique<AndroidSurfaceMock>();
+  auto surface_factory = std::make_shared<TestAndroidSurfaceFactory>(
+      [&android_context, gr_context, window]() {
+        auto android_surface_mock =
+            std::make_unique<AndroidSurfaceMock>(android_context);
         EXPECT_CALL(*android_surface_mock, CreateGPUSurface(gr_context.get()));
         EXPECT_CALL(*android_surface_mock, SetNativeWindow(window));
         EXPECT_CALL(*android_surface_mock, IsValid()).WillOnce(Return(true));
@@ -81,9 +82,10 @@ TEST(SurfacePool, GetUnusedLayers) {
           ByMove(std::make_unique<PlatformViewAndroidJNI::OverlayMetadata>(
               0, window))));
 
-  auto surface_factory =
-      std::make_shared<TestAndroidSurfaceFactory>([gr_context, window]() {
-        auto android_surface_mock = std::make_unique<AndroidSurfaceMock>();
+  auto surface_factory = std::make_shared<TestAndroidSurfaceFactory>(
+      [&android_context, gr_context, window]() {
+        auto android_surface_mock =
+            std::make_unique<AndroidSurfaceMock>(android_context);
         EXPECT_CALL(*android_surface_mock, CreateGPUSurface(gr_context.get()));
         EXPECT_CALL(*android_surface_mock, SetNativeWindow(window));
         EXPECT_CALL(*android_surface_mock, IsValid()).WillOnce(Return(true));
@@ -116,8 +118,9 @@ TEST(SurfacePool, GetLayerRecycle) {
 
   auto gr_context_2 = GrDirectContext::MakeMock(nullptr);
   auto surface_factory = std::make_shared<TestAndroidSurfaceFactory>(
-      [gr_context_1, gr_context_2, window]() {
-        auto android_surface_mock = std::make_unique<AndroidSurfaceMock>();
+      [&android_context, gr_context_1, gr_context_2, window]() {
+        auto android_surface_mock =
+            std::make_unique<AndroidSurfaceMock>(android_context);
         // Allocate two GPU surfaces for each gr context.
         EXPECT_CALL(*android_surface_mock,
                     CreateGPUSurface(gr_context_1.get()));
@@ -163,9 +166,10 @@ TEST(SurfacePool, GetLayerAllocateTwoLayers) {
           ByMove(std::make_unique<PlatformViewAndroidJNI::OverlayMetadata>(
               1, window))));
 
-  auto surface_factory =
-      std::make_shared<TestAndroidSurfaceFactory>([gr_context, window]() {
-        auto android_surface_mock = std::make_unique<AndroidSurfaceMock>();
+  auto surface_factory = std::make_shared<TestAndroidSurfaceFactory>(
+      [&android_context, gr_context, window]() {
+        auto android_surface_mock =
+            std::make_unique<AndroidSurfaceMock>(android_context);
         EXPECT_CALL(*android_surface_mock, CreateGPUSurface(gr_context.get()));
         EXPECT_CALL(*android_surface_mock, SetNativeWindow(window));
         EXPECT_CALL(*android_surface_mock, IsValid()).WillOnce(Return(true));
@@ -202,9 +206,10 @@ TEST(SurfacePool, DestroyLayers) {
           ByMove(std::make_unique<PlatformViewAndroidJNI::OverlayMetadata>(
               0, window))));
 
-  auto surface_factory =
-      std::make_shared<TestAndroidSurfaceFactory>([gr_context, window]() {
-        auto android_surface_mock = std::make_unique<AndroidSurfaceMock>();
+  auto surface_factory = std::make_shared<TestAndroidSurfaceFactory>(
+      [&android_context, gr_context, window]() {
+        auto android_surface_mock =
+            std::make_unique<AndroidSurfaceMock>(android_context);
         EXPECT_CALL(*android_surface_mock, CreateGPUSurface(gr_context.get()));
         EXPECT_CALL(*android_surface_mock, SetNativeWindow(window));
         EXPECT_CALL(*android_surface_mock, IsValid()).WillOnce(Return(true));
@@ -231,9 +236,10 @@ TEST(SurfacePool, DestroyLayersFrameSizeChanged) {
 
   auto window = fml::MakeRefCounted<AndroidNativeWindow>(nullptr);
 
-  auto surface_factory =
-      std::make_shared<TestAndroidSurfaceFactory>([gr_context, window]() {
-        auto android_surface_mock = std::make_unique<AndroidSurfaceMock>();
+  auto surface_factory = std::make_shared<TestAndroidSurfaceFactory>(
+      [&android_context, gr_context, window]() {
+        auto android_surface_mock =
+            std::make_unique<AndroidSurfaceMock>(android_context);
         EXPECT_CALL(*android_surface_mock, CreateGPUSurface(gr_context.get()));
         EXPECT_CALL(*android_surface_mock, SetNativeWindow(window));
         EXPECT_CALL(*android_surface_mock, IsValid()).WillOnce(Return(true));
