+++ shell/platform/android/platform_view_android.cc
@@ -31,29 +31,36 @@ namespace flutter {
 
 AndroidSurfaceFactoryImpl::AndroidSurfaceFactoryImpl(
     const std::shared_ptr<AndroidContext>& context,
-    bool enable_impeller)
-    : android_context_(context), enable_impeller_(enable_impeller) {}
+    std::shared_ptr<PlatformViewAndroidJNI> jni_facade,
+    bool enable_impeller,
+    bool enable_vulkan_validation)
+    : android_context_(context),
+      jni_facade_(std::move(jni_facade)),
+      enable_impeller_(enable_impeller),
+      enable_vulkan_validation_(enable_vulkan_validation) {}
 
 AndroidSurfaceFactoryImpl::~AndroidSurfaceFactoryImpl() = default;
 
 std::unique_ptr<AndroidSurface> AndroidSurfaceFactoryImpl::CreateSurface() {
   switch (android_context_->RenderingApi()) {
     case AndroidRenderingAPI::kSoftware:
-      return std::make_unique<AndroidSurfaceSoftware>();
+      return std::make_unique<AndroidSurfaceSoftware>(android_context_,
+                                                      jni_facade_);
     case AndroidRenderingAPI::kOpenGLES:
       if (enable_impeller_) {
-        return std::make_unique<AndroidSurfaceGLImpeller>(
-            std::static_pointer_cast<AndroidContextGLImpeller>(
-                android_context_));
+// TODO(kaushikiska@): Enable this after wiring a preference for Vulkan backend.
+#if false
+        return std::make_unique<AndroidSurfaceVulkanImpeller>(
+            android_context_, jni_facade_, enable_vulkan_validation_);
+
+#else
+        return std::make_unique<AndroidSurfaceGLImpeller>(android_context_,
+                                                          jni_facade_);
+#endif
       } else {
-        return std::make_unique<AndroidSurfaceGLSkia>(
-            std::static_pointer_cast<AndroidContextGLSkia>(android_context_));
+        return std::make_unique<AndroidSurfaceGLSkia>(android_context_,
+                                                      jni_facade_);
       }
-    case AndroidRenderingAPI::kVulkan:
-      FML_DCHECK(enable_impeller_);
-      return std::make_unique<AndroidSurfaceVulkanImpeller>(
-          std::static_pointer_cast<AndroidContextVulkanImpeller>(
-              android_context_));
     default:
       FML_DCHECK(false);
       return nullptr;
@@ -64,19 +71,12 @@ static std::shared_ptr<flutter::AndroidContext> CreateAndroidContext(
     bool use_software_rendering,
     const flutter::TaskRunners& task_runners,
     uint8_t msaa_samples,
-    bool enable_impeller,
-    bool enable_vulkan_validation) {
+    bool enable_impeller) {
   if (use_software_rendering) {
     return std::make_shared<AndroidContext>(AndroidRenderingAPI::kSoftware);
   }
   if (enable_impeller) {
-    // TODO(kaushikiska@): Enable this after wiring a preference for Vulkan
-    // backend.
-#if false
-    return std::make_unique<AndroidContextVulkanImpeller>(enable_vulkan_validation);
-#else
     return std::make_unique<AndroidContextGLImpeller>();
-#endif
   }
   return std::make_unique<AndroidContextGLSkia>(
       AndroidRenderingAPI::kOpenGLES,               //
@@ -100,8 +100,7 @@ PlatformViewAndroid::PlatformViewAndroid(
               use_software_rendering,
               task_runners,
               msaa_samples,
-              delegate.OnPlatformViewGetSettings().enable_impeller,
-              delegate.OnPlatformViewGetSettings().enable_vulkan_validation)) {}
+              delegate.OnPlatformViewGetSettings().enable_impeller)) {}
 
 PlatformViewAndroid::PlatformViewAndroid(
     PlatformView::Delegate& delegate,
@@ -117,8 +116,10 @@ PlatformViewAndroid::PlatformViewAndroid(
     FML_CHECK(android_context_->IsValid())
         << "Could not create surface from invalid Android context.";
     surface_factory_ = std::make_shared<AndroidSurfaceFactoryImpl>(
-        android_context_,                                     //
-        delegate.OnPlatformViewGetSettings().enable_impeller  //
+        android_context_,                                              //
+        jni_facade_,                                                   //
+        delegate.OnPlatformViewGetSettings().enable_impeller,          //
+        delegate.OnPlatformViewGetSettings().enable_vulkan_validation  //
     );
     android_surface_ = surface_factory_->CreateSurface();
     FML_CHECK(android_surface_ && android_surface_->IsValid())
