+++ shell/platform/android/platform_view_android.h
@@ -27,7 +27,9 @@ namespace flutter {
 class AndroidSurfaceFactoryImpl : public AndroidSurfaceFactory {
  public:
   AndroidSurfaceFactoryImpl(const std::shared_ptr<AndroidContext>& context,
-                            bool enable_impeller);
+                            std::shared_ptr<PlatformViewAndroidJNI> jni_facade,
+                            bool enable_impeller,
+                            bool enable_vulkan_validation);
 
   ~AndroidSurfaceFactoryImpl() override;
 
@@ -35,7 +37,9 @@ class AndroidSurfaceFactoryImpl : public AndroidSurfaceFactory {
 
  private:
   const std::shared_ptr<AndroidContext>& android_context_;
+  std::shared_ptr<PlatformViewAndroidJNI> jni_facade_;
   const bool enable_impeller_;
+  const bool enable_vulkan_validation_;
 };
 
 class PlatformViewAndroid final : public PlatformView {
