+++ shell/platform/android/surface/android_surface.cc
@@ -7,7 +7,11 @@
 
 namespace flutter {
 
-AndroidSurface::AndroidSurface() = default;
+AndroidSurface::AndroidSurface(
+    const std::shared_ptr<AndroidContext>& android_context) {
+  FML_DCHECK(android_context->IsValid());
+  android_context_ = android_context;
+}
 
 AndroidSurface::~AndroidSurface() = default;
