+++ shell/platform/android/surface/android_surface.h
@@ -46,7 +46,9 @@ class AndroidSurface {
   virtual std::shared_ptr<impeller::Context> GetImpellerContext();
 
  protected:
-  AndroidSurface();
+  explicit AndroidSurface(
+      const std::shared_ptr<AndroidContext>& android_context);
+  std::shared_ptr<AndroidContext> android_context_;
 };
 
 class AndroidSurfaceFactory {
