+++ shell/platform/android/surface/android_surface_mock.cc
@@ -6,6 +6,10 @@
 
 namespace flutter {
 
+AndroidSurfaceMock::AndroidSurfaceMock(
+    const std::shared_ptr<AndroidContext>& android_context)
+    : AndroidSurface(android_context) {}
+
 std::unique_ptr<GLContextResult> AndroidSurfaceMock::GLContextMakeCurrent() {
   return std::make_unique<GLContextDefaultResult>(/*static_result=*/true);
 }
